requirements.txt
jupyter
pandas
numpy
matplotlib
tqdm
biopython
pathlib
requests
